+++ packages/react-tinacms-inline/src/blocks/inline-field-blocks.tsx
@@ -40,6 +40,8 @@ export interface InlineBlocksProps {
   itemProps?: {
     [key: string]: any
   }
+  min?: number
+  max?: number
 }
 
 export interface InlineBlocksActions {
@@ -53,6 +55,8 @@ export interface InlineBlocksActions {
   activeBlock: number | null
   setActiveBlock: any
   direction: 'vertical' | 'horizontal'
+  min?: number
+  max?: number
 }
 
 export const InlineBlocksContext = React.createContext<InlineBlocksActions | null>(
@@ -75,6 +79,8 @@ export function InlineBlocks({
   className,
   direction = 'vertical',
   itemProps,
+  min,
+  max,
 }: InlineBlocksProps) {
   const cms = useCMS()
   const [activeBlock, setActiveBlock] = useState(-1)
@@ -128,6 +134,8 @@ export function InlineBlocks({
                       activeBlock,
                       setActiveBlock,
                       direction,
+                      min,
+                      max,
                     }}
                   >
                     {allData.length < 1 && cms.enabled && (