+++ packages/@tinacms/forms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.2](https://github.com/tinacms/tinacms/compare/@tinacms/forms@0.6.1...@tinacms/forms@0.6.2) (2020-06-29)
+
+**Note:** Version bump only for package @tinacms/forms
+
+
+
+
+
 ## [0.6.1](https://github.com/tinacms/tinacms/compare/@tinacms/forms@0.6.1-alpha.0...@tinacms/forms@0.6.1) (2020-06-23)
 
 **Note:** Version bump only for package @tinacms/forms